+++ cxroots/contours/annulus_sector.py
@@ -78,7 +78,7 @@ class AnnulusSector(Contour):
             ComplexArc(center, r0, phi1, phi0 - phi1),
         ]
 
-        super(AnnulusSector, self).__init__(segments)
+        super().__init__(segments)
 
     def __str__(self):
         return (